build(ruff): migrate to 0.16 and clear the expanded default rule set (#631)
* build(ruff): adopt 0.16 defaults and pin CI to 0.16.0

Ruff 0.16 enables ~413 rules by default (up from 59). The repo had no ruff
config at all, so it inherited whatever the defaults were.

- Add [tool.ruff] with required-version = ">=0.16.0" so a stale local ruff
  fails loudly rather than disagreeing with CI.
- Ignore BLE001 (blind-except; semantic pass tracked in CODE-REVIEW.md),
  I001 (facade import order is a deliberate DAG), DTZ005/DTZ007 (local
  wall-clock timestamps), UP047 (PEP 695 generics).
- per-file-ignores for tests/**: B018 (stream-consuming expressions),
  RUF012 (mutable attrs on test doubles).
- Bump the CI lint job pin 0.15.12 -> 0.16.0; sync CLOUD.md and the check skill
  (ruff format now covers Markdown code blocks too).

* style: apply ruff 0.16 safe autofixes

224 mechanical fixes from the newly-default rule set:

- RUF100: drop 59 noqa directives for rules that no longer (or never did) fire.
  The six that carried justification prose keep it as a plain comment.
- UP009: drop 42 redundant `# -*- coding: utf-8 -*-` headers.
- UP017: datetime.timezone.utc -> datetime.UTC (32).
- UP035: typing.Callable/Iterator -> collections.abc (21), hand-placed in sorted
  position since I001 stays off.
- PIE807 lambda: [] -> list, PLR1711/RET501 useless returns, PLR1730 min/max,
  FURB167 re.I -> re.IGNORECASE, UP037 unquoted annotations, and assorted singles.

466 -> 106 findings. Full suite green (2238 passed).

* refactor: apply reviewed ruff 0.16 unsafe autofixes

75 fixes, applied and read rule by rule rather than in one blanket pass:

- RUF046 (32): drop the redundant int() around single-arg round(), which
  already returns int (verified on numpy scalars too, not just floats).
- ISC004 (9): parenthesize implicit string concatenation inside list literals.
  All nine were intentional line wraps, not missing commas.
- C408 dict() -> {}, RUF059 unused unpacked vars -> _, PIE810 startswith tuples,
  RUF007 zip(x, x[1:]) -> itertools.pairwise, SIM118 .keys() drops, UP031/UP032
  %-format -> f-string, UP022 capture_output=True, plus assorted singles.

Hand-cleaned two autofix artifacts: a redundant bool() wrapper in
ollama_client._is_connection_refused and redundant parens in titlecards.

106 -> 33 findings. Full suite green (2238 passed).

* fix: resolve remaining ruff 0.16 lint findings

The 33 findings with no autofix, decided case by case:

- PLW1510 (7, video.py): explicit check=False. Every ffmpeg/ffprobe call here
  inspects returncode itself; check=True would raise past that handling.
- SIM102 (9): collapsed 5 nested ifs; kept 4 nested with a reason — two where an
  explanatory comment sits between the branches, two in spreadsheet.py where the
  row-presence/bounds split matches the neighbouring severity_cell guard.
- S110 (6): narrowed two resp.close() handlers to OSError and one to ImportError,
  logged the workflow progress-notify failure via verbose_print, and left the
  studio rollback pair swallowing with a noqa (the existing comment explains why).
- TRY004 (2): kept ValueError with a noqa. Every caller catches ValueError to
  render a hint or a 400; TypeError would escape those and 500 instead.
- PLW0602 (4): dropped globals that are only read.
- SIM115: tempfile now opened via a with-block, so a failing write can't leak the
  handle; delete=False and the explicit unlink stay (ffmpeg reopens it by path).
- PERF402 append-loop -> extend, PLC0206 dict iteration -> .items().

ruff check is clean. Full suite green (2238 passed).

* style: format python code blocks in markdown

Ruff 0.16 formats Python fences in .md by default. Reflows snippets in
AGENTS.md and three archived plans; the AGENTS.md clip-record block keeps its
one-line-per-field shape (comment shortened so the formatter leaves it inline).

+++ screenspace_primitives.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """Screenspace image-analysis primitives (pure cv2/numpy).
 
 Region cropping/denormalization/resolution, HSV color math, frame-diff, SSIM,
@@ -10,7 +9,8 @@ buffer, static-frame skip). No file or ffmpeg I/O lives here.
 import functools
 import math
 import statistics
-from typing import TYPE_CHECKING, Any, Callable
+from collections.abc import Callable
+from typing import TYPE_CHECKING, Any
 
 import cv2
 
@@ -18,7 +18,7 @@ try:
     # cv2.data.haarcascades is a submodule attribute; ty needs the explicit
     # import. Absent on cv2 builds without the bundled cascade data (the face
     # channel feature-detects it and degrades to a zeros map).
-    import cv2.data  # noqa: F401
+    import cv2.data
 except ImportError:  # pragma: no cover - depends on the installed cv2 build
     pass
 import numpy as np
@@ -306,10 +306,10 @@ def denormalize_region(
         ``mask_points``/``shape`` passed through for shaped regions.
     """
     out: dict[str, Any] = {
-        "x": int(round(region["x"] * target_w)),
-        "y": int(round(region["y"] * target_h)),
-        "w": int(round(region["w"] * target_w)),
-        "h": int(round(region["h"] * target_h)),
+        "x": round(region["x"] * target_w),
+        "y": round(region["y"] * target_h),
+        "w": round(region["w"] * target_w),
+        "h": round(region["h"] * target_h),
     }
     # Shaped regions: contour vertices are bbox-relative (0-1 of the region's
     # own rect), so they pass through denormalization verbatim. Copying them
@@ -366,7 +366,10 @@ def resolve_region_request(
         raise ValueError(f"Region '{region_name}' not found")
 
     if not isinstance(region_ref, dict):
-        raise ValueError("region_ref must be an object")
+        # ValueError, not TypeError: this resolver's contract is "ValueError means
+        # bad request", and every caller (cli.py, screenspace_server's
+        # _resolve_region_request) catches ValueError to render a hint or a 400.
+        raise ValueError("region_ref must be an object")  # noqa: TRY004
 
     source = str(region_ref.get("source", "")).strip()
 
@@ -672,8 +675,8 @@ def _scale_template(
     if not (effective > 0 and abs(effective - 1.0) > 1e-6):
         return template, mask
     th, tw = template.shape[:2]
-    nw = max(8, int(round(tw * effective)))
-    nh = max(8, int(round(th * effective)))
+    nw = max(8, round(tw * effective))
+    nh = max(8, round(th * effective))
     interp = cv2.INTER_AREA if effective < 1.0 else cv2.INTER_CUBIC
     scaled_template = cv2.resize(template, (nw, nh), interpolation=interp)
     scaled_mask = (
